nba_service/apps/core/views.py
"""Core views — health check."""

from rest_framework.request import Request
from rest_framework.response import Response
from rest_framework.views import APIView


class HealthCheckView(APIView):
    """Simple health check endpoint."""

    def get(self, request: Request) -> Response:  # noqa: ARG002
        return Response({"status": "ok"})
